Removed composer hash checking from installation
When the hash changes, installations break until I get the hash
updated. I’m not sure if the security is warranty in a docker container
in which no sensitive information likely ever exists. There may however
be a better way to verify against a URL or something, but that might
not be any more secure (since it could be spoofed).